add : ColumnController, ColumnService, ColumnRepository, ColumnRequestDto, ColumnResponseDto, update : ColumnController, Service (컬럼 생성 기능 구현), Errorcode(예외 케이스 추가)

+++ src/main/java/com/sparta/sensorypeople/domain/column/entity/ColumnResponseDto.java
@@ -8,7 +8,7 @@ public class ColumnResponseDto {
 
     private String ColumnName;
 
-    private int ColumnOrder;
+    private Double ColumnOrder;
 
     public ColumnResponseDto(Columns column) {
         this.id=column.getId();
